% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primal_dual_saddle.R
\name{find_saddle_point}
\alias{find_saddle_point}
\title{Title}
\usage{
find_saddle_point(
  im_vec,
  mask_vec,
  task,
  lmda,
  u0,
  tol,
  max_iter,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
