% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restorer.R
\name{tv_restore_primal_dual}
\alias{tv_restore_primal_dual}
\title{Title}
\usage{
tv_restore_primal_dual(
  img,
  mask,
  lambda,
  tau,
  sigma,
  theta,
  max_iter,
  tol,
  verbose
)
}
\arguments{
\item{theta}{}
}
\description{
Title
}
