% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{plot_images}
\alias{plot_images}
\title{plot_images
Plot one or more grayscale images in a grid with titles}
\usage{
plot_images(
  imgs,
  titles = NULL,
  cols = NULL,
  palette = grey(seq(0, 1, length.out = 256)),
  useRaster = TRUE,
  axes = FALSE,
  main = NULL
)
}
\arguments{
\item{imgs}{A list of image matrices (rows × cols) for grayscale images.}

\item{titles}{Character vector of same length as \code{imgs}. Titles to display above each image. If \code{NULL}, no individual titles.}

\item{cols}{Integer or \code{NULL}. Number of columns in the grid layout (by row-major order). If \code{NULL}, it uses \code{floor(sqrt(length(imgs)))} columns.}

\item{palette}{Character vector. Colour ramp for grayscale (default: \code{grey(seq(0,1,length.out=256))}).}

\item{useRaster}{Logical. Whether to use raster graphics (default: \code{TRUE}).}

\item{axes}{Logical. Whether to draw axes around each image (default: \code{FALSE}).}

\item{main}{Character scalar. Overall main title for the combined grid (default: \code{NULL}).}
}
\description{
plot_images
Plot one or more grayscale images in a grid with titles
}
\examples{
plot_images()
}
