% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restorer.R
\name{tv_restore}
\alias{tv_restore}
\title{Title}
\usage{
tv_restore(
  img,
  mask = NULL,
  lambda,
  method = "primal_dual",
  task = "ROF_inpainting",
  u0 = NULL,
  tau,
  sigma,
  theta,
  max_iter = 300,
  tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{theta}{}
}
\description{
Title
}
